move check_architecture to distribute_utils
